Reconstruct the overlay 5 asset row lookup
An asset row names the image and palette entries for a graphic and can select
an alternate row for the current save file. The lookup reuses a resident
registry entry or builds a reference-counted asset and issues the loads.

Matching C/C++ reaches 324,864 of 1,563,700 bytes (20.78%).

+++ src/overlay005/asset_loader.c
@@ -116,7 +116,7 @@ void func_ov005_02066dbc(Overlay5Archive *archive, u16 group, u16 index,
     owner->image = entry;
 }
 
-void func_ov005_02066d04(Overlay5Archive *archive, u8 index, u16 entry,
+void func_ov005_02066d04(Overlay5Archive *archive, u16 index, u16 entry,
                          Overlay5BufferEntry *owner, u8 mode, int heap) {
     Overlay5ArchiveRequest *request = func_ov005_02067184(archive, owner->key);
     ArchiveCompressedRequest *io;
